requirements.txt
annoy==1.12.0
aspell-python-py3==1.15
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
Django==2.0.6
django-cors-headers==2.2.0
faiss==0.1
fasteners==0.14.1
future==0.16.0
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
lz4==2.0.0
monotonic==1.5
numpy==1.14.4
pandas==0.23.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pymagnitude==0.1.32
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
torch==0.4.0
torchtext==0.2.3
tqdm==4.23.4
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
xxhash==1.0.1
